// libs/foundation/dashboard/src/lib/admin-dashboard/models/engagement-details.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText } from '@gosi-ui/core';
export class EngagementDetails {
  count: number = undefined;
  nationality: BilingualText = new BilingualText();
}